+++ src/Codeception/Module/Laravel5.php
@@ -528,12 +528,11 @@ class Laravel5 extends Framework implements ActiveRecord, PartedModule
      */
     public function haveRecord($tableName, $attributes = [])
     {
-        $id = $this->app['db']->table($tableName)->insertGetId($attributes);
-        if (!$id) {
-            $this->fail("Couldn't insert record into table $tableName");
+        try {
+            return $this->app['db']->table($tableName)->insertGetId($attributes);
+        } catch (\Exception $e) {
+            $this->fail("Couldn't insert record into table $tableName: " . $e->getMessage());
         }
-
-        return $id;
     }
 
     /**